tests: use traits to test owned and borrowed types
This lets us ensure that the behaviour of Root/Handle and
RootRef/HandleRef is consistent (because unfortunately we can't use
Deref so technically the two types have separate implementations and so
we could very easily end up with one of the types having a bug due to a
copy-paste issue).

By defining a trait, we can also prove a capi-backed trait impl that
will let us test the C API from within our regular test suite. We can
also make sure that methods are added to both types (and eventually the
C API) because we cannot write a test using the new method without
addding it to the trait and adding an implementation.
